requirements.txt
asgiref==3.6.0
charset-normalizer==3.1.0
coinbase-commerce==1.0.1
Django==3.2.18
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==2.0.2
wincertstore==0.2
